requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
scipy
opencv-python
matplotlib
tqdm
Pillow
dlib
insightface
onnxruntime
onnxruntime-gpu
openai-clip
open_clip_torch